feat(back): create backend to App
- create models to database
- create API

P.S. I know that it is not customary to make commits in such large parts, and I know that there are many points in the code that should be redone. But right now it's more important for me that it works at all. I'll deal with the problems later. May be